server/requirements.txt
pillow==9.5.0
numpy==1.24.0
requests==2.28.2
opencv-python-headless==4.7.0.72
python-multipart==0.0.5  # Add this line
flask==2.3.2